Adjust chatbot example on re-run (#137)
* Some preparation for re-running chatbot experiments

* Modify space

* Use the exhaustive optimizer

* Fix to cacheing

* Add logging to cache

* Skip cohere

* Modify logging

* Fix reporting

* Skip failed runs

* Remove Alpaca

* Adjust number of systems

+++ zeno_build/cache_utils.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 import hashlib
 import json
+import logging
 import os
 import pathlib
 from typing import Any
@@ -14,16 +15,37 @@ def get_cache_path(
     params: dict[str, Any],
     extension: str | None = None,
 ) -> str:
-    """Get a path to a cache.
+    """Get the path to a cache.
 
     Args:
-        task: The task to cache for.
+        cache_root: The root of the cache path.
         params: The parameters that the cache should represent.
         extension: The extension to use for the cache file, or None for no
             extension.
 
     Returns:
         The path to the cache file or directory.
+    """
+    _, cache_path = get_cache_id_and_path(cache_root, params, extension)
+    return cache_path
+
+
+def get_cache_id_and_path(
+    cache_root: str,
+    params: dict[str, Any],
+    extension: str | None = None,
+) -> tuple[str, str]:
+    """Get the ID and path to a cache.
+
+    Args:
+        cache_root: The root of the cache path.
+        params: The parameters that the cache should represent.
+        extension: The extension to use for the cache file, or None for no
+            extension.
+
+    Returns:
+        - The cache ID, which is the hash of the parameters.
+        - The path to the cache file or directory.
     """
     if extension == "zbp":
         raise ValueError(
@@ -45,7 +67,7 @@ def get_cache_path(
                 break
     with open(param_file, "w") as f:
         f.write(dumped_params)
-    return os.path.join(
+    return base_name, os.path.join(
         cache_root, base_name if extension is None else f"{base_name}.{extension}"
     )
 
@@ -63,6 +85,8 @@ class CacheLock:
     def __init__(self, filename: str):
         """Initialize the lock."""
         self.lock_path = f"{filename}.zblock"
+        self.fail_path = f"{filename}.zbfail"
+        self.skipped = False
 
     def __enter__(self) -> bool:
         """Enter the cache lock.
@@ -71,15 +95,20 @@ class CacheLock:
         Otherwise, return False.
         """
         # Skip if the lock file exists
-        if os.path.exists(self.lock_path):
+        if os.path.exists(self.lock_path) or os.path.exists(self.fail_path):
+            logging.getLogger(__name__).debug(f"Skipping {self.lock_path}")
+            self.skipped = True
             return False
 
         # Create the lock file
         with open(self.lock_path, "w"):
+            logging.getLogger(__name__).debug(f"Created lock file {self.lock_path}")
             pass
 
         return True
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         """Remove the lock file."""
-        os.remove(self.lock_path)
+        if not self.skipped:
+            logging.getLogger(__name__).debug(f"Deleting lock {self.lock_path}")
+            os.remove(self.lock_path)
